Updated float to CGFloat.

+++ ResearchKit/Common/ORKHTMLPDFWriter.m
@@ -36,10 +36,10 @@
 #define PPI 72
 #define ORKSizeMakeWithPPI(width, height) CGSizeMake(width * PPI, height * PPI)
 
-static const float A4Width = 8.26666667;
-static const float A4Height = 11.6916667;
-static const float LetterWidth = 8.5f;
-static const float LetterHeight = 11.0f;
+static const CGFloat A4Width = 8.26666667;
+static const CGFloat A4Height = 11.6916667;
+static const CGFloat LetterWidth = 8.5f;
+static const CGFloat LetterHeight = 11.0f;
 
 
 #pragma mark - ORKHTMLPDFWriter Interface
